+++ contracts/cron/StrategyManagerStorage.sol
@@ -3,10 +3,10 @@
 pragma solidity >=0.7.2;
 pragma experimental ABIEncoderV2;
 
-import "../admin/AdminableInit.sol";
-import "../interface/IUniswapV2Router02.sol";
-import "@openzeppelin/contracts/math/SafeMath.sol";
-import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import '../admin/AdminableInit.sol';
+import '../interface/IUniswapV2Router02.sol';
+import '@openzeppelin/contracts/math/SafeMath.sol';
+import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 
 contract StrategyManagerStorage is AdminableInit {
     using SafeMath for uint256;
@@ -23,10 +23,15 @@ contract StrategyManagerStorage is AdminableInit {
     function initialize(address _store) public initializer {
         AdminableInit.initializeAdmin(_store);
     }
-    
-    function setPath(address _from, address _to, address[] memory _path, address _router) external onlyAdmin {
-        require(_from == _path[0], "The first token of the Uniswap route must be the from token"); 
-        require(_to == _path[_path.length - 1], "The last token of the Uniswap route must be the to token");
+
+    function setPath(
+        address _from,
+        address _to,
+        address[] memory _path,
+        address _router
+    ) external onlyAdmin {
+        require(_from == _path[0], 'The first token of the Uniswap route must be the from token');
+        require(_to == _path[_path.length - 1], 'The last token of the Uniswap route must be the to token');
         PathItem memory item;
         item.path = _path;
         item.router = _router;
@@ -41,14 +46,17 @@ contract StrategyManagerStorage is AdminableInit {
         return pathes[_from][_to];
     }
 
-    function swap(address _fromToken, address _toToken, uint amount) external returns (uint) {
-        uint balanceBefore = IERC20(_fromToken).balanceOf(address(this));
+    function swap(
+        address _fromToken,
+        address _toToken,
+        uint256 amount
+    ) external returns (uint256) {
+        uint256 balanceBefore = IERC20(_fromToken).balanceOf(address(this));
         IERC20(_fromToken).transferFrom(msg.sender, address(this), amount);
-        uint balanceAfter = IERC20(_fromToken).balanceOf(address(this));
-        require(balanceAfter.sub(balanceBefore) == amount, "illegal token transfer");
+        uint256 balanceAfter = IERC20(_fromToken).balanceOf(address(this));
+        require(balanceAfter.sub(balanceBefore) == amount, 'illegal token transfer');
         PathItem memory item = pathFor(_fromToken, _toToken);
-        uint[] memory amounts = IUniswapV2Router02(item.router).swapExactTokensForTokens(amount, 0, item.path, msg.sender, block.timestamp);
+        uint256[] memory amounts = IUniswapV2Router02(item.router).swapExactTokensForTokens(amount, 0, item.path, msg.sender, block.timestamp);
         return amounts[amounts.length - 1];
     }
-
 }
